Rename account folder to wallet
A wallet contains multiple accounts, and eventually we'll have multiple
wallets.
