+++ ironfish/src/index.ts
@@ -1,7 +1,7 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-export * from './account'
+export * from './wallet'
 export * from './assert'
 export * from './blockchain'
 export * from './consensus'
