+++ ironfish/src/indexers/minedBlocksIndexer.ts
@@ -1,7 +1,6 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { Accounts } from '../account'
 import { Assert } from '../assert'
 import { Blockchain } from '../blockchain'
 import { ChainProcessor } from '../chainProcessor'
@@ -18,6 +17,7 @@ import {
 import { createDB } from '../storage/utils'
 import { SetTimeoutToken } from '../utils'
 import { BlockchainUtils, isBlockMine } from '../utils/blockchain'
+import { Accounts } from '../wallet'
 import {
   AccountsToRemoveValue,
   AccountsToRemoveValueEncoding,
