+++ ironfish/src/node.ts
@@ -3,7 +3,6 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import os from 'os'
 import { v4 as uuid } from 'uuid'
-import { Accounts, AccountsDB } from './account'
 import { Blockchain } from './blockchain'
 import {
   Config,
@@ -27,6 +26,7 @@ import { RpcServer } from './rpc/server'
 import { Strategy } from './strategy'
 import { Syncer } from './syncer'
 import { Telemetry } from './telemetry/telemetry'
+import { Accounts, AccountsDB } from './wallet'
 import { WorkerPool } from './workerPool'
 
 export class IronfishNode {
