+++ ironfish/src/rpc/routes/accounts/utils.ts
@@ -1,8 +1,8 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import { Account } from '../../../account'
 import { IronfishNode } from '../../../node'
+import { Account } from '../../../wallet'
 import { ValidationError } from '../../adapters'
 
 export function getAccount(node: IronfishNode, name?: string): Account {
