+++ ironfish/src/sdk.test.ts
@@ -2,13 +2,13 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 import os from 'os'
-import { Accounts } from './account'
 import { Config, DEFAULT_DATA_DIR } from './fileStores'
 import { NodeFileProvider } from './fileSystems'
 import { IronfishNode } from './node'
 import { Platform } from './platform'
 import { RpcClient, RpcMemoryClient } from './rpc'
 import { IronfishSdk } from './sdk'
+import { Accounts } from './wallet'
 
 describe('IronfishSdk', () => {
   describe('init', () => {
