+++ ironfish/src/testUtilities/fixtures.ts
@@ -4,13 +4,13 @@
 import { generateKey } from '@ironfish/rust-nodejs'
 import fs from 'fs'
 import path from 'path'
-import { Account, Accounts, AccountsValue } from '../account'
 import { Assert } from '../assert'
 import { Blockchain } from '../blockchain'
 import { IronfishNode } from '../node'
 import { Block, BlockSerde, SerializedBlock } from '../primitives/block'
 import { SerializedTransaction, Transaction } from '../primitives/transaction'
 import { IJSON } from '../serde'
+import { Account, Accounts, AccountsValue } from '../wallet'
 import { getCurrentTestPath } from './utils'
 
 const FIXTURE_FOLDER = '__fixtures__'
