+++ ironfish/src/testUtilities/nodeTest.ts
@@ -4,7 +4,6 @@
 import os from 'os'
 import path from 'path'
 import { v4 as uuid } from 'uuid'
-import { Accounts } from '../account'
 import { Blockchain } from '../blockchain'
 import { Verifier } from '../consensus/verifier'
 import { ConfigOptions } from '../fileStores/config'
@@ -12,6 +11,7 @@ import { PeerNetwork } from '../network'
 import { IronfishNode } from '../node'
 import { IronfishSdk } from '../sdk'
 import { Syncer } from '../syncer'
+import { Accounts } from '../wallet'
 import { WorkerPool } from '../workerPool'
 import { TestStrategy } from './strategy'
