+++ ironfish/src/wallet/validator.ts
